@@ -1,16 +1,16 @@
 {
   "name": "medic-conf",
-  "version": "2.2.0",
+  "version": "3.0.6",
   "description": "Configure Medic Mobile deployments",
-  "main": "index.js",
+  "main": "./src/lib/main.js",
   "engines": {
-    "node": ">=8.9.0"
+    "node": ">=8.10.0"
   },
   "scripts": {
     "clean": "rm -rf ./build/",
     "generate-changelog": "echo 'Generating changelog...' && ./scripts/changelog > CHANGELOG.md && echo 'Changelog generated.'",
     "eslint": "eslint src/**/*.js test/*.js test/**/*.js",
-    "test": "npm run eslint && npm run clean && mkdir -p build/test && cp -r test/data build/test/data && cd build/test && mocha ../../test/**/*.spec.js",
+    "test": "npm run eslint && npm run clean && mkdir -p build/test && cp -r test/data build/test/data && cd build/test && mocha ../../test/**/*.spec.js && cd ../.. && npx @medic/audit-dependencies check",
     "postversion": "npm publish",
     "postpublish": "git push --all; git push --tags"
   },
@@ -32,35 +32,35 @@
   "homepage": "https://github.com/medic/medic-conf#readme",
   "dependencies": {
     "canonical-json": "0.0.4",
-    "csv-parse": "^3.1.3",
-    "googleapis": "^34.0.0",
-    "jshint": "^2.9.6",
-    "json2csv": "^4.3.5",
+    "csv-parse": "^4.4.3",
+    "eslint": "^6.5.0",
+    "eslint-loader": "^3.0.2",
+    "googleapis": "^43.0.0",
+    "iso-639-1": "^2.1.0",
+    "json2csv": "^4.5.1",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
     "opn": "^6.0.0",
-    "pouchdb-adapter-http": "^7.0.0",
-    "pouchdb-core": "^7.0.0",
-    "pouchdb-mapreduce": "^7.0.0",
-    "readline-sync": "^1.4.9",
+    "pouchdb-adapter-http": "^7.1.1",
+    "pouchdb-core": "^7.1.1",
+    "pouchdb-mapreduce": "^7.1.1",
+    "readline-sync": "^1.4.10",
     "redact-basic-auth": "^1.0.0",
     "request": "^2.88.0",
-    "request-promise-native": "^1.0.5",
-    "semver": "^5.6.0",
-    "svgo": "^1.1.1",
-    "uglify-js": "^3.5.0",
-    "uuid": "^3.3.2"
+    "request-promise-native": "^1.0.7",
+    "semver": "^6.1.1",
+    "svgo": "^1.2.2",
+    "uuid": "^3.3.3",
+    "webpack": "^4.41.0"
   },
   "devDependencies": {
     "chai": "^4.2.0",
-    "chai-exclude": "^1.0.10",
+    "chai-exclude": "^2.0.1",
     "chai-shallow-deep-equal": "^1.4.6",
-    "eslint": "^5.15.1",
-    "express-pouchdb": "^4.1.0",
-    "lodash": "^4.17.11",
-    "mocha": "^5.2.0",
-    "pouchdb-adapter-memory": "^7.0.0",
+    "express-pouchdb": "^4.2.0",
+    "mocha": "^6.1.4",
+    "pouchdb-adapter-memory": "^7.1.1",
     "rewire": "^4.0.1",
-    "sinon": "^7.3.1"
+    "sinon": "^7.3.2"
   }
 }
