@@ -12,7 +12,8 @@ module.exports = shell => {
   } else if (shell === true){
     console.log('# ERROR shell type argument not provided e.g. --shell-completion=bash');
   } else {
-    console.log('# ERROR medic-conf shell completion not yet supported for', shell);
+    console.error('shell completion not yet supported for', shell);
+    process.exit(1);
   }
   process.exit(1);
 };
