@@ -3,25 +3,20 @@ const supportedActions = require('./supported-actions');
 module.exports = () => {
   console.log(`
 ${bold('NAME')}
-  medic-conf - configurerize your Medic Mobile instances
+  medic-conf - Configure your Medic Mobile instances
 
 ${bold('SYNOPSIS')}
   medic-conf <--local|--instance=instance-name|--url=url>
 Or:
-  medic-conf <--local|--instance=instance-name|--url=url> <actions> -- <params>
+  medic-conf <--local|--instance=instance-name|--url=url|--archive> <actions> -- <params>
 
 ${bold('DESCRIPTION')}
-  This script will update and upload all of a particular project
-  configuration to a particular instance.
-
-  Assumptions are made about the layout of content for a project, and these
-  should not be configurable - the script is only intended for use by
-  properly-structured projects.
+  This script updates and uploads a project's configuration.
 
 ${bold('SUPPORTED ACTIONS')}
   * ${supportedActions.join('\n  * ')}
 
-${bold('SPECIFYING URL')}
+${bold('SAVE CONFIG TO')}
   --local
     Upload to http://admin:pass@localhost:5988
 
@@ -34,6 +29,9 @@ ${bold('SPECIFYING URL')}
   --url=<url>
     Upload to URL specified.
 
+  --archive
+    Save configuration content to files on disk. Requires parameter --destination=<path to save files>.
+
 ${bold('OPTIONS')}
   --help
     Display this usage message
@@ -41,6 +39,8 @@ ${bold('OPTIONS')}
   --shell-completion
   --shell-completion=bash
     Generate the shell-completion script for use in bash.
+
+  --source=<path to project folder>. Defaults to the working directory.
 
   --supported-actions
     Display a list of supported actions.
@@ -50,9 +50,12 @@ ${bold('OPTIONS')}
 
   --changelog
     Display application changelog.
-  
+
   --accept-self-signed-certs
-    Allows medic-conf to work with self signed certs by telling node to ignore the error 
+    Allows medic-conf to work with self signed certs by telling node to ignore the error
+
+  --skip-dependency-check
+    Skips checking the version running is set to the same version in the package.json
 `);
 };
 
