@@ -1,9 +1,10 @@
 const compressPng = require('../lib/compress-png');
 const compressSvg = require('../lib/compress-svg');
+const environment = require('../lib/environment');
 const fs = require('../lib/sync-fs');
 
-module.exports = (projectDir/*, couchUrl*/) =>
-  fs.recurseFiles(projectDir)
+module.exports = () =>
+  fs.recurseFiles(environment.pathToProject)
     .reduce((promiseChain, path) => {
         switch(fs.extension(path)) {
           case 'png': return compressPng(promiseChain, path);
