@@ -1,5 +1,6 @@
 const convertForms = require('../lib/convert-forms');
+const environment = require('../lib/environment');
 
-module.exports = (projectDir, couchUrl, extras) => convertForms(projectDir, 'collect', {
-  forms: extras,
+module.exports = () => convertForms(environment.pathToProject, 'collect', {
+  forms: environment.extraArgs,
 });
