@@ -1,12 +1,9 @@
-const skipFn = require('../lib/skip-fn');
+const formsList = require('../lib/forms-list');
 const pouch = require('../lib/db');
 
-module.exports = (projectDir, couchUrl) => {
-  if(!couchUrl) return skipFn('no couch URL set');
-
-  const db = pouch(couchUrl);
-
-  return db.getForms({ include_docs:true })
+module.exports = () => {
+  const db = pouch();
+  return formsList(db, { include_docs: true })
     .then(res => res.rows)
     .then(forms => Promise.all(forms.map(f => db.remove(f.doc))));
 };
