@@ -1,5 +1,6 @@
+const environment = require('../lib/environment');
 const uploadForms = require('../lib/upload-forms');
 
-module.exports = (projectDir, couchUrl, extras) => uploadForms(projectDir, couchUrl, 'app', {
-  forms: extras,
+module.exports = () => uploadForms(environment.pathToProject, 'app', {
+  forms: environment.extraArgs,
 });
