@@ -1,17 +1,10 @@
+const api = require('../lib/api');
+const environment = require('../lib/environment');
 const fs = require('../lib/sync-fs');
-const request = require('request-promise-native');
-const skipFn = require('../lib/skip-fn');
 
-module.exports = (projectDir, couchUrl) => {
-  if(!couchUrl) return skipFn('no couch URL set');
-
-  return request
-    .put({
-      method: 'PUT',
-      url: `${couchUrl}/_design/medic/_rewrite/update_settings/medic?replace=1`,
-      headers: { 'Content-Type':'application/json' },
-      body: fs.read(`${projectDir}/app_settings.json`),
-    })
+module.exports = () => {
+  const settings = fs.read(`${environment.pathToProject}/app_settings.json`);
+  return api().updateAppSettings(settings)
     .then(JSON.parse)
     .then(json => {
       // As per https://github.com/medic/medic-webapp/issues/3674, this endpoint
