@@ -1,6 +1,7 @@
+const environment = require('../lib/environment');
 const uploadForms = require('../lib/upload-forms');
 
-module.exports = (projectDir, couchUrl, extras) => uploadForms(projectDir, couchUrl, 'collect', {
-  forms: extras,
+module.exports = () => uploadForms(environment.pathToProject, 'collect', {
+  forms: environment.extraArgs,
   default_context: { collect:true },
 });
