@@ -1,18 +1,18 @@
-const fs = require('../lib/sync-fs');
-const skipFn = require('../lib/skip-fn');
-const warn = require('../lib/log').warn;
-const pouch = require('../lib/db');
-const request = require('request-promise-native');
 const semver = require('semver');
+
+const environment = require('../lib/environment');
+const fs = require('../lib/sync-fs');
+const pouch = require('../lib/db');
+const getApiVersion = require('../lib/get-api-version');
+const iso639 = require('iso-639-1');
+const { warn } = require('../lib/log');
 
 const FILE_MATCHER = /messages-.*\.properties/;
 
-module.exports = (projectDir, couchUrl) => {
-  if(!couchUrl) return skipFn('no couch URL set');
+module.exports = () => {
+  const db = pouch(environment.apiUrl);
 
-  const dir = `${projectDir}/translations`;
-  const db = pouch(couchUrl);
-  const instanceUrl = couchUrl.replace(/\/medic$/, '');
+  const dir = `${environment.pathToProject}/translations`;
 
   return Promise.resolve()
     .then(() => {
@@ -21,12 +21,28 @@ module.exports = (projectDir, couchUrl) => {
       return Promise.all(fs.readdir(dir)
         .filter(name => FILE_MATCHER.test(name))
         .map(fileName => {
+          const id = idFor(fileName);
+          const languageCode = id.substring('messages-'.length);
+          let languageName = iso639.getName(languageCode);
+          if (!languageName){
+            warn(`'${languageCode}' is not a recognized ISO 639 language code, please ask admin to set the name`);
+            languageName = 'TODO: please ask admin to set this in settings UI';
+          } else {
+            let languageNativeName = iso639.getNativeName(languageCode);
+            if (languageNativeName !== languageName){
+              languageName = `${languageNativeName} (${languageName})`;
+            }
+          }
+
           var translations = propertiesAsObject(`${dir}/${fileName}`);
 
-          return db.get(idFor(fileName))
+          return db.get(id)
             .catch(e => {
-              if(e.status === 404) return newDocFor(fileName, instanceUrl, db);
-              else throw e;
+              if(e.status === 404) {
+                return newDocFor(fileName, db, languageName, languageCode);
+              }
+              
+              throw e;
             })
             .then(doc => overwriteProperties(doc, translations))
             .then(doc => db.put(doc));
@@ -51,10 +67,8 @@ function overwriteProperties(doc, props) {
     doc.custom = props;
   } else if (doc.values) {
     // pre-3.4.0 doc structure
-    for(const k in props) {
-      if(props.hasOwnProperty(k)) {
-        doc.values[k] = props[k];
-      }
+    for (const [key, value] of Object.entries(props)) {
+      doc.values[key] = value;
     }
   } else {
     throw new Error(`Existent translation doc ${doc._id} is malformed`);
@@ -63,47 +77,37 @@ function overwriteProperties(doc, props) {
   return doc;
 }
 
-function newDocFor(fileName, instanceUrl, db) {
-  const id = idFor(fileName);
-
+async function newDocFor(fileName, db, languageName, languageCode) {
   const doc = {
-    _id: id,
+    _id: idFor(fileName),
     type: 'translations',
-    code: id.substring(id.indexOf('-') + 1),
-    name: 'TODO: please ask admin to set this in settings UI',
+    code: languageCode,
+    name: languageName,
     enabled: true,
   };
 
-  return genericTranslationsStructure(instanceUrl, db).then(useGenericTranslations => {
-    if (useGenericTranslations) {
-      doc.generic = {};
-    } else {
-      doc.values = {};
-    }
+  const useGenericTranslations = await genericTranslationsStructure(db);
+  if (useGenericTranslations) {
+    doc.generic = {};
+  } else {
+    doc.values = {};
+  }
 
-    return doc;
-  });
+  return doc;
 }
 
 function idFor(fileName) {
   return fileName.substring(0, fileName.length - 11);
 }
 
-const getVersion = (instanceUrl, db) => {
-  return request({ uri: `${instanceUrl}/api/deploy-info`, method: 'GET', json: true }) // endpoint added in 3.5
-    .catch(() => db.get('_design/medic-client').then(doc => doc.deploy_info)) // since 3.0.0
-    .then(deploy_info => deploy_info && deploy_info.version);
-};
+async function genericTranslationsStructure(db) {
+  const version = await getApiVersion();
 
-const genericTranslationsStructure = (instanceUrl, db) => {
-  return getVersion(instanceUrl, db).then(version => {
-    if (semver.valid(version)) {
-      return semver.gte(version, '3.4.0');
-    }
+  if (semver.valid(version)) {
+    return semver.gte(version, '3.4.0');
+  }
 
-    return db
-      .get('messages-en')
-      .then(doc => doc.generic)
-      .catch(() => false);
-  });
-};
+  return db.get('messages-en')
+    .then(doc => doc.generic)
+    .catch(() => false);
+}
