@@ -1,17 +1,27 @@
 const { assert, expect } = require('chai');
-const compileAppSettings = require('../../src/fn/compile-app-settings');
+const path = require('path');
+const sinon = require('sinon');
+const rewire = require('rewire');
+
+const compileAppSettings = rewire('../../src/fn/compile-app-settings');
+const environment = require('../../src/lib/environment');
 const fs = require('../../src/lib/sync-fs');
 
+let writeJson;
 describe('compile-app-settings', () => {
+  beforeEach(() => {
+    writeJson = sinon.stub(fs, 'writeJson');
+    compileAppSettings.__set__('fs', fs);
+  });
+  afterEach(() => {
+    sinon.restore();
+  });
 
   it('should handle simple config', () =>
     test('simple/project'));
 
   it('should handle derivative app-settings definitions', () =>
     test('derivative/child'));
-
-  it('should handle nools & contact-summary templating', () =>
-    test('templating/project'));
 
   it('should handle config with no separate task-schedules.json file', () =>
     test('no-task-schedules.json/project'));
@@ -22,39 +32,67 @@ describe('compile-app-settings', () => {
   it('should reject a project with both old and new nools config', () =>
     testFails('unexpected-legacy-nools-rules/project'));
 
-  it('should handle a project with a purge function', () =>
-    test('purging-function/project'));
+  it('should reject a project with both purge and purging files', () =>
+    testFails('purge/both-purge-and-purging/project'));
 
-  it('should handle a project with a perge function that need to be merged with other purge config', () =>
-    test('purging-function/project'));
+  it('should reject a project purge file exists and its not valid js', () =>
+    testFails('purge/broken-purge-file/project'));
+
+  it('should reject a project invalid purge file config', () =>
+    testFails('purge/invalid-purge/project'));
 
   it('should reject a project with an uncompilable purging function', () =>
-    testFails('invalid-purging-function/project'));
+    testFails('purge/invalid-purging-function/project'));
+
+  it('should handle a project with a purge function that need to be merged with other purge config', () =>
+    test('purge/merge-purging-function/project'));
+
+  it('should handle a project with no export purge config', () =>
+    test('purge/no-export-purge/project'));
+
+  it('should handle a project with correct purge config', () =>
+    test('purge/purge-correct/project'));
+
+  it('should reject a project where purge.fn is not a function', () =>
+    testFails('purge/purge-fn-not-a-function/project'));
+
+  it('should handle a project with a purge function', () =>
+    test('purge/purging-function/project'));
+
+  it('should reject a project with eslint error', () =>
+    testFails('eslint-error/project'));
+
+  it('can overwrite eslint rules with eslintrc file', () =>
+    test('eslintrc/project'));
 });
 
-function test(relativeProjectDir) {
-  const testDir = `./data/compile-app-settings/${relativeProjectDir}`;
+async function test(relativeProjectDir) {
+  const testDir = path.join(__dirname, '../data/compile-app-settings', relativeProjectDir);
+  sinon.stub(environment, 'pathToProject').get(() => testDir);
 
   // when
-  return compileAppSettings(testDir)
+  await compileAppSettings();
 
-    .then(() => {
-      // then
-      const actual = JSON.parse(fs.read(`${testDir}/app_settings.json`));
-      const expected = JSON.parse(fs.read(`${testDir}/../app_settings.expected.json`));
-      actual.tasks.rules = expected.tasks.rules = '';
-      expect(actual).to.deep.eq(expected);
-    });
+  // then
+  const actual = JSON.parse(JSON.stringify(writeJson.args[0][1]));
+  const expected = JSON.parse(fs.read(`${testDir}/../app_settings.expected.json`));
+  actual.tasks.rules = expected.tasks.rules = '';
+  actual.contact_summary = expected.contact_summary = '';
+  expect(actual).to.deep.eq(expected);
 }
 
-function testFails(relativeProjectDir) {
-  const testDir = `./data/compile-app-settings/${relativeProjectDir}`;
+async function testFails(relativeProjectDir) {
+  const testDir = path.join(__dirname, '../data/compile-app-settings', relativeProjectDir);
 
   // when
-  return compileAppSettings(testDir)
-    .then(() => assert.fail('Expected compileAppSettings() to fail, but it didn\'t.'))
-    .catch(e => {
-      if(e.name === 'AssertionError') throw e;
-      assert.ok(e);
-    });
+  try {
+    await compileAppSettings(testDir);
+    assert.fail('Expected assertion');
+  } catch (err) {
+    if (err.name === 'AssertionError') {
+      throw err;
+    }
+
+    assert.ok('asserted');
+  }
 }
