@@ -1,13 +1,15 @@
-const assert = require('chai').assert;
-const fs = require('../../src/lib/sync-fs');
+const { assert } = require('chai');
+const sinon = require('sinon');
 
+const environment = require('../../src/lib/environment');
+const fs = require('../../src/lib/sync-fs');
 
 module.exports = {
   testFor: (testName, type) => {
 
     const convertForms = require(`../../src/fn/convert-${type}-forms`);
 
-    describe(testName, function() {
+    describe(testName, function () {
 
       this.timeout(30000); // allow time for form conversion
 
@@ -16,24 +18,23 @@ module.exports = {
       // recursively copy forms and expected XML to temp directory, and create
       // tests dynamically
 
-      const targetDir = `${projectDir}/forms/${type}`;
+      const expectedDir = `${projectDir}/forms/${type}/expected`;
 
-      fs.recurseFiles(targetDir)
-        .filter(file => file.endsWith('.expected.xml'))
-        .forEach(expectedXml => {
+      fs.recurseFiles(expectedDir).forEach(expectedXml => {
 
-          const generatedXml = expectedXml.replace(/\.expected\.xml$/, '.xml');
+        const generatedXml = expectedXml.replace('/expected/', '/');
 
-          it(`should generate ${generatedXml} as expected`, () => {
-
-            assert.ok(fs.exists(generatedXml), `Missing generated XML file: ${generatedXml}`);
-            assert.equal(fs.read(generatedXml), fs.read(expectedXml), `Content of ${generatedXml} was not as expected.`);
-
-          });
-
+        it(`should generate ${generatedXml} as expected`, () => {
+          assert.ok(fs.exists(generatedXml), `Missing generated XML file: ${generatedXml}`);
+          assert.equal(fs.read(generatedXml), fs.read(expectedXml), `Content of ${generatedXml} was not as expected.`);
         });
 
-      before(() => convertForms(projectDir));
+      });
+
+      before(() => {
+        sinon.stub(environment, 'pathToProject').get(() => projectDir);
+        return convertForms();
+      });
 
     });
 
