@@ -1,7 +1,9 @@
-const assert = require('chai').assert;
-const fs = require('../../src/lib/sync-fs');
+const { assert } = require('chai');
+const sinon = require('sinon');
 
 const csvToDocs = require('../../src/fn/csv-to-docs');
+const environment = require('../../src/lib/environment');
+const fs = require('../../src/lib/sync-fs');
 
 describe('csv-to-docs', function() {
   this.timeout(30000); // allow time for slow things
@@ -14,9 +16,10 @@ describe('csv-to-docs', function() {
       it(`should convert demo files in ${dir} to expected JSON`, function(done) {
         // given
         dir = `${testDir}/${dir}`;
+        sinon.stub(environment, 'pathToProject').get(() => dir);
 
         // when
-        csvToDocs(dir)
+        csvToDocs()
           .then(() => {
             const generatedDocsDir = `${dir}/json_docs`;
             const expectedDocsDir  = `${dir}/expected-json_docs`;
