@@ -7,7 +7,6 @@ PouchDB.plugin(require('pouchdb-mapreduce'));
 
 const moveContactsModule = rewire('../../src/fn/move-contacts');
 moveContactsModule.__set__('prepareDocumentDirectory', () => {});
-moveContactsModule.__set__('console', { log: () => {} });
 const updateLineagesAndStage = moveContactsModule.__get__('updateLineagesAndStage');
 const { mockReport, mockHierarchy, parentsToLineage } = require('../mock-hierarchies');
 
@@ -296,7 +295,7 @@ describe('move-contacts', () => {
 
     await upsert('clinic_1', clinic);
     await upsert('patient_1', patient);
-    
+
     await updateLineagesAndStage({
       contactIds: ['clinic_1'],
       parentId: 'district_2',
