@@ -1,7 +1,12 @@
+const { assert } = require('chai');
+const sinon = require('sinon');
+
 const api = require('../api-stub');
-const assert = require('chai').assert;
-const uploadCustomTranslations = require('../../src/fn/upload-custom-translations');
+const environment = require('../../src/lib/environment');
 const testProjectDir = './data/upload-custom-translations/';
+const uploadCustomTranslations = require('../../src/fn/upload-custom-translations');
+
+const mockTestDir = testDir => sinon.stub(environment, 'pathToProject').get(() => `${testProjectDir}${testDir}`);
 
 const getTranslationDoc = (lang) => {
   return api.db.get(`messages-${lang}`)
@@ -35,7 +40,8 @@ describe('upload-custom-translations', () => {
     });
 
     it('should upload simple translations', () => {
-      return uploadCustomTranslations(`${testProjectDir}simple`, api.couchUrl)
+      mockTestDir(`simple`);
+      return uploadCustomTranslations()
         .then(() => expectTranslationDocs('en'))
         .then(() => getTranslationDoc('en'))
         .then(messagesEn => {
@@ -46,16 +52,19 @@ describe('upload-custom-translations', () => {
     });
 
     it('should upload translations for multiple languages', () => {
-      return uploadCustomTranslations(`${testProjectDir}multi-lang`, api.couchUrl)
+      mockTestDir(`multi-lang`);
+      return uploadCustomTranslations()
         .then(() => expectTranslationDocs('en', 'fr'))
         .then(() => getTranslationDoc('en'))
         .then(messagesEn => {
+          assert(messagesEn.name === 'English');
           assert.deepEqual(messagesEn.values, { one: 'one' });
           assert(!messagesEn.generic);
           assert(!messagesEn.custom);
         })
         .then(() => getTranslationDoc('fr'))
         .then(messagesFr => {
+          assert(messagesFr.name === 'Français (French)');
           assert.deepEqual(messagesFr.values, { one: 'un(e)' });
           assert(!messagesFr.generic);
           assert(!messagesFr.custom);
@@ -63,7 +72,8 @@ describe('upload-custom-translations', () => {
     });
 
     it('should upload translations containing equals signs', () => {
-      return uploadCustomTranslations(`${testProjectDir}contains-equals`, api.couchUrl)
+      mockTestDir(`contains-equals`);
+      return uploadCustomTranslations()
         .then(() => expectTranslationDocs('en'))
         .then(() => getTranslationDoc('en'))
         .then(messagesEn => {
@@ -77,6 +87,7 @@ describe('upload-custom-translations', () => {
     });
 
     it('should work correctly when falling back to testing messages-en', () => {
+      mockTestDir(`custom-lang`);
       return api.db
         .put({
           _id: 'messages-en',
@@ -85,7 +96,7 @@ describe('upload-custom-translations', () => {
           type: 'translations',
           values: { a: 'first' }
         })
-        .then(() => uploadCustomTranslations(`${testProjectDir}custom-lang`, api.couchUrl))
+        .then(() => uploadCustomTranslations())
         .then(() => expectTranslationDocs('en', 'fr'))
         .then(() => getTranslationDoc('en'))
         .then(messagesEn => {
@@ -100,6 +111,16 @@ describe('upload-custom-translations', () => {
           assert(!messagesFr.custom);
         });
     });
+
+    it('should set default name for unknown language', () => {
+      mockTestDir(`unknown-lang`);
+      return uploadCustomTranslations()
+        .then(() => expectTranslationDocs('qp'))
+        .then(() => getTranslationDoc('qp'))
+        .then(messagesQp => {
+          assert(messagesQp.name === 'TODO: please ask admin to set this in settings UI');
+        });
+    });
   });
 
   describe('medic-3.x', () => {
@@ -109,7 +130,8 @@ describe('upload-custom-translations', () => {
       it('should upload simple translations', () => {
         // api/deploy-info endpoint doesn't exist
         api.giveResponses({ status: 404, body: { error: 'not_found' } });
-        return uploadCustomTranslations(`${testProjectDir}simple`, api.couchUrl)
+        mockTestDir(`simple`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -122,16 +144,19 @@ describe('upload-custom-translations', () => {
       it('should upload translations for multiple languages', () => {
         // api/deploy-info endpoint doesn't exist
         api.giveResponses({ status: 404, body: { error: 'not_found' } });
-        return uploadCustomTranslations(`${testProjectDir}multi-lang`, api.couchUrl)
+        mockTestDir(`multi-lang`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en', 'fr'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
+            assert(messagesEn.name === 'English');
             assert.deepEqual(messagesEn.values, { one: 'one' });
             assert(!messagesEn.generic);
             assert(!messagesEn.custom);
           })
           .then(() => getTranslationDoc('fr'))
           .then(messagesFr => {
+            assert(messagesFr.name === 'Français (French)');
             assert.deepEqual(messagesFr.values, { one: 'un(e)' });
             assert(!messagesFr.generic);
             assert(!messagesFr.custom);
@@ -141,7 +166,8 @@ describe('upload-custom-translations', () => {
       it('should upload translations containing equals signs', () => {
         // api/deploy-info endpoint doesn't exist
         api.giveResponses({ status: 404, body: { error: 'not_found' } });
-        return uploadCustomTranslations(`${testProjectDir}contains-equals`, api.couchUrl)
+        mockTestDir(`contains-equals`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -155,6 +181,7 @@ describe('upload-custom-translations', () => {
       });
 
       it('should merge with existent translations', () => {
+        mockTestDir(`with-customs`);
         return api.db
           .put({
             _id: 'messages-en',
@@ -163,7 +190,7 @@ describe('upload-custom-translations', () => {
             type: 'translations',
             values: { a:'first', from_custom:'third' }
           })
-          .then(() => uploadCustomTranslations(`${testProjectDir}with-customs`, api.couchUrl))
+          .then(() => uploadCustomTranslations())
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -174,6 +201,7 @@ describe('upload-custom-translations', () => {
       });
 
       it('should crash for malformed translation files', () => {
+        mockTestDir(`with-customs`);
         return api.db
           .put({
             _id: 'messages-en',
@@ -181,9 +209,19 @@ describe('upload-custom-translations', () => {
             name: 'English',
             type: 'translations'
           })
-          .then(() => uploadCustomTranslations(`${testProjectDir}with-customs`, api.couchUrl))
+          .then(() => uploadCustomTranslations())
           .catch(err => {
             assert.equal(err.message, 'Existent translation doc messages-en is malformed');
+          });
+      });
+
+      it('should set default name for unknown language', () => {
+        mockTestDir(`unknown-lang`);
+        return uploadCustomTranslations()
+          .then(() => expectTranslationDocs('qp'))
+          .then(() => getTranslationDoc('qp'))
+          .then(messagesQp => {
+            assert(messagesQp.name === 'TODO: please ask admin to set this in settings UI');
           });
       });
     });
@@ -194,7 +232,8 @@ describe('upload-custom-translations', () => {
       it('should upload simple translations', () => {
         // api/deploy-info endpoint doesn't exist
         api.giveResponses({ status: 404, body: { error: 'not_found' } });
-        return uploadCustomTranslations(`${testProjectDir}simple`, api.couchUrl)
+        mockTestDir(`simple`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -207,16 +246,19 @@ describe('upload-custom-translations', () => {
       it('should upload translations for multiple languages', () => {
         // api/deploy-info endpoint doesn't exist
         api.giveResponses({ status: 404, body: { error: 'not_found' } });
-        return uploadCustomTranslations(`${testProjectDir}multi-lang`, api.couchUrl)
+        mockTestDir(`multi-lang`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en', 'fr'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
+            assert(messagesEn.name === 'English');
             assert.deepEqual(messagesEn.custom, { one: 'one' });
             assert.deepEqual(messagesEn.generic, {});
             assert(!messagesEn.values);
           })
           .then(() => getTranslationDoc('fr'))
           .then(messagesFr => {
+            assert(messagesFr.name === 'Français (French)');
             assert.deepEqual(messagesFr.custom, { one: 'un(e)' });
             assert.deepEqual(messagesFr.generic, {});
             assert(!messagesFr.values);
@@ -226,7 +268,8 @@ describe('upload-custom-translations', () => {
       it('should upload translations containing equals signs', () => {
         // api/deploy-info endpoint doesn't exist
         api.giveResponses({ status: 404, body: { error: 'not_found' } });
-        return uploadCustomTranslations(`${testProjectDir}contains-equals`, api.couchUrl)
+        mockTestDir(`contains-equals`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -240,6 +283,7 @@ describe('upload-custom-translations', () => {
       });
 
       it('should replace existent custom values', () => {
+        mockTestDir(`with-customs`);
         return api.db
           .put({
             _id: 'messages-en',
@@ -249,7 +293,7 @@ describe('upload-custom-translations', () => {
             generic: { a: 'first' },
             custom: { c: 'third' }
           })
-          .then(() => uploadCustomTranslations(`${testProjectDir}with-customs`, api.couchUrl))
+          .then(() => uploadCustomTranslations())
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -260,6 +304,7 @@ describe('upload-custom-translations', () => {
       });
 
       it('should replace delete custom values', () => {
+        mockTestDir(`no-customs`);
         return api.db
           .put({
             _id: 'messages-en',
@@ -269,7 +314,7 @@ describe('upload-custom-translations', () => {
             generic: { a: 'first' },
             custom: { c: 'third' }
           })
-          .then(() => uploadCustomTranslations(`${testProjectDir}no-customs`, api.couchUrl))
+          .then(() => uploadCustomTranslations())
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -282,6 +327,7 @@ describe('upload-custom-translations', () => {
       it('should work correctly when falling back to testing messages-en', () => {
         // api/deploy-info endpoint doesn't exist
         api.giveResponses({ status: 404, body: { error: 'not_found' } });
+        mockTestDir(`custom-lang`);
         // for *some* reason, medic-client doesn't have deploy-info
         return api.db
           .get('_design/medic-client')
@@ -296,7 +342,7 @@ describe('upload-custom-translations', () => {
             type: 'translations',
             generic: { a: 'first' }
           }))
-          .then(() => uploadCustomTranslations(`${testProjectDir}custom-lang`, api.couchUrl))
+          .then(() => uploadCustomTranslations())
           .then(() => expectTranslationDocs('en', 'fr'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -311,6 +357,16 @@ describe('upload-custom-translations', () => {
             assert(!messagesFr.values);
           });
       });
+
+      it('should set default name for unknown language', () => {
+        mockTestDir(`unknown-lang`);
+        return uploadCustomTranslations()
+          .then(() => expectTranslationDocs('qp'))
+          .then(() => getTranslationDoc('qp'))
+          .then(messagesQp => {
+            assert(messagesQp.name === 'TODO: please ask admin to set this in settings UI');
+          });
+      });
     });
 
     describe('3.5.0', () => {
@@ -321,7 +377,8 @@ describe('upload-custom-translations', () => {
       });
 
       it('should upload simple translations', () => {
-        return uploadCustomTranslations(`${testProjectDir}simple`, api.couchUrl)
+        mockTestDir(`simple`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -332,16 +389,19 @@ describe('upload-custom-translations', () => {
       });
 
       it('should upload translations for multiple languages', () => {
-        return uploadCustomTranslations(`${testProjectDir}multi-lang`, api.couchUrl)
+        mockTestDir(`multi-lang`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en', 'fr'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
+            assert(messagesEn.name === 'English');
             assert.deepEqual(messagesEn.custom, { one: 'one' });
             assert.deepEqual(messagesEn.generic, {});
             assert(!messagesEn.values);
           })
           .then(() => getTranslationDoc('fr'))
           .then(messagesFr => {
+            assert(messagesFr.name === 'Français (French)');
             assert.deepEqual(messagesFr.custom, { one: 'un(e)' });
             assert.deepEqual(messagesFr.generic, {});
             assert(!messagesFr.values);
@@ -349,7 +409,8 @@ describe('upload-custom-translations', () => {
       });
 
       it('should upload translations containing equals signs', () => {
-        return uploadCustomTranslations(`${testProjectDir}contains-equals`, api.couchUrl)
+        mockTestDir(`contains-equals`);
+        return uploadCustomTranslations()
           .then(() => expectTranslationDocs('en'))
           .then(() => getTranslationDoc('en'))
           .then(messagesEn => {
@@ -361,6 +422,19 @@ describe('upload-custom-translations', () => {
             assert(!messagesEn.values);
           });
       });
+
+      it('should set default name for unknown language', () => {
+        mockTestDir(`unknown-lang`);
+        return uploadCustomTranslations()
+          .then(() => expectTranslationDocs('qp'))
+          .then(() => getTranslationDoc('qp'))
+          .then(messagesQp => {
+            assert(messagesQp.name === 'TODO: please ask admin to set this in settings UI');
+          });
+      });
+
     });
+
+
   });
 });
