@@ -23,7 +23,7 @@ describe('upload-docs', function() {
 
   it('should upload docs to pouch', async () => {
     await assertDbEmpty();
-    await uploadDocs('', api.couchUrl);
+    await uploadDocs();
     const res = await api.db.allDocs();
 
     expect(res.rows.map(doc => doc.id)).to.deep.eq(['one', 'three', 'two']);
@@ -31,11 +31,11 @@ describe('upload-docs', function() {
 
   it('do nothing if there are no docs to upload', async () => {
     await assertDbEmpty();
-    
+
     const pouch = sinon.stub();
     fs.recurseFiles = () => [];
     return uploadDocs.__with__({ fs, pouch })(async () => {
-      await uploadDocs('', api.couchUrl);
+      await uploadDocs();
       expect(pouch.called).to.be.false;
     });
   });
@@ -48,7 +48,7 @@ describe('upload-docs', function() {
 
     return uploadDocs.__with__({ fs, pouch })(async () => {
       try {
-        await uploadDocs('', api.couchUrl);
+        await uploadDocs();
         expect.fail('should throw');
       } catch (err) {
         expect(err.message).to.include('expected _id is');
@@ -62,14 +62,14 @@ describe('upload-docs', function() {
       .returns(Promise.resolve([{}]));
     fs.recurseFiles = () => new Array(10).fill('').map((x, i) => `${i}.doc.json`);
     sinon.useFakeTimers(0);
-    
+
     const imported_date = new Date(0).toISOString();
     return uploadDocs.__with__({
       INITIAL_BATCH_SIZE: 4,
       fs,
       pouch: () => ({ bulkDocs }),
     })(async () => {
-      await uploadDocs('', api.couchUrl);
+      await uploadDocs();
       expect(bulkDocs.callCount).to.eq(1 + 10 / 2);
 
       // first failed batch of 4
