@@ -2,24 +2,27 @@ const rewire = require('rewire');
 const sinon = require('sinon');
 const { expect } = require('chai');
 
+const environment = rewire('../../src/lib/environment');
 const main = rewire('../../src/lib/main');
-const normalArgv = ['node', 'medic-conf'];
 
 const defaultActions = main.__get__('defaultActions');
+const normalArgv = ['node', 'medic-conf'];
 
 let mocks;
 describe('main', () => {
   beforeEach(() => {
+    environment.__set__('state', {});
+    sinon.spy(environment, 'initialize');
     mocks = {
       usage: sinon.stub(),
-      console: {
-        log: sinon.stub(),
-      },
       shellCompletionSetup: sinon.stub(),
       error: sinon.stub(),
       info: sinon.stub(),
+      checkMedicConfDependencyVersion: sinon.stub(),
       warn: sinon.stub(),
       executeAction: sinon.stub(),
+      getApiUrl: sinon.stub().returns('http://api'),
+      environment,
       readline: {
         question: sinon.stub().returns('pwd'),
         keyInYN: sinon.stub().returns(true),
@@ -35,6 +38,9 @@ describe('main', () => {
     for (let attr of Object.keys(mocks)) {
       main.__set__(attr, mocks[attr]);
     }
+  });
+  afterEach(() => {
+    environment.initialize.restore();
   });
 
   it('no argv yields usage', async () => {
@@ -55,14 +61,24 @@ describe('main', () => {
 
   it('--supported-actions', async () => {
     await main([...normalArgv, '--supported-actions'], {});
-    expect(mocks.console.log.callCount).to.eq(1);
-    expect(mocks.console.log.args[0]).to.include('Supported actions:\n');
+    expect(mocks.info.callCount).to.eq(1);
+    expect(mocks.info.args[0]).to.include('Supported actions:\n');
   });
 
   it('--version', async () => {
     await main([...normalArgv, '--version'], {});
-    expect(mocks.console.log.callCount).to.eq(1);
-    expect(mocks.console.log.args[0]).to.match(/[0-9]+\.[0-9]+\.[0-9]/);
+    expect(mocks.info.callCount).to.eq(1);
+    expect(mocks.info.args[0]).to.match(/[0-9]+\.[0-9]+\.[0-9]/);
+  });
+
+  it('--skip-dependency-check', async () => {
+    await main([...normalArgv, '--skip-dependency-check'], {});
+    expect(mocks.checkMedicConfDependencyVersion.callCount).to.eq(0);
+  });
+
+  it('medic conf dependency checked', async () => {
+    await main([...normalArgv, '--local'], {});
+    expect(mocks.checkMedicConfDependencyVersion.calledOnce).to.be.true;
   });
 
   it('--local --accept-self-signed-certs', async () => {
@@ -71,48 +87,54 @@ describe('main', () => {
     expect(main.__get__('process').env.NODE_TLS_REJECT_UNAUTHORIZED).to.eq('0');
   });
 
+  it('errors if you do not provide an instance when required', async () => {
+    mocks.getApiUrl.returns();
+
+    await main([...normalArgv, 'backup-all-forms'], {});
+
+    expect(mocks.executeAction.called).to.be.false;
+  });
+
+  it('supports actions that do not require an instance', async () => {
+    await main([...normalArgv, 'initialise-project-layout'], {});
+    expect(mocks.executeAction.callCount).to.deep.eq(1);
+    expect(mocks.executeAction.args[0][0].name).to.eq('initialise-project-layout');
+  });
+
+  const expectExecuteActionBehavior = (expectedActions, expectedExtraParams) => {
+    if (Array.isArray(expectedActions)) {
+      expect(mocks.executeAction.args.map(args => args[0].name)).to.deep.eq(expectedActions);
+    } else {
+      expect(mocks.executeAction.args[0][0].name).to.eq(expectedActions);
+    }
+
+    expect(mocks.environment.initialize.args[0][3]).to.deep.eq(expectedExtraParams);
+    expect(mocks.environment.initialize.args[0][4]).to.eq('http://api');
+  };
+
   it('--local no COUCH_URL', async () => {
     await main([...normalArgv, '--local'], {});
-    
-    expect(mocks.executeAction.callCount).to.deep.eq(defaultActions.length);
-    expect(mocks.executeAction.args[0]).to.deep.eq(['compile-app-settings', 'http://admin:pass@localhost:5988/medic', undefined]);
+    expectExecuteActionBehavior(defaultActions, undefined);
   });
 
   it('--local with COUCH_URL to localhost', async () => {
     const COUCH_URL = 'http://user:pwd@localhost:5988/medic';
     await main([...normalArgv, '--local'], { COUCH_URL });
-    
-    expect(mocks.executeAction.callCount).to.deep.eq(defaultActions.length);
-    expect(mocks.executeAction.args[0]).to.deep.eq(['compile-app-settings', 'http://user:pwd@localhost:5988/medic', undefined]);
-  });
-
-  it('--local with COUCH_URL to non-localhost yields error', async () => {
-    const COUCH_URL = 'http://user:pwd@host:5988/medic';
-    await main([...normalArgv, '--local'], { COUCH_URL });
-    expect(mocks.executeAction.callCount).to.deep.eq(0);
-  });
-
-  it('--instance production warning', async () => {
-    mocks.readline.keyInYN.returns(false);
-    const exit = await main([...normalArgv, '--instance=resolved.app', '--user=foo'], {});
-    expect(mocks.readline.question.calledOnce).to.be.true; // prompt for password
-    expect(mocks.readline.keyInYN.calledOnce).to.be.true; // prompted with warning
-    expect(mocks.warn.args[0][1]).to.include('https://foo:****@resolved.app.medicmobile.org');
-    expect(exit).to.eq(-1);
+    expectExecuteActionBehavior(defaultActions, undefined);
   });
 
   it('--instance + 2 ordered actions', async () => {
     await main([...normalArgv, '--instance=test.app', 'convert-app-forms', 'compile-app-settings'], {});
     expect(mocks.executeAction.callCount).to.deep.eq(2);
-    expect(mocks.executeAction.args[0]).to.deep.eq(['convert-app-forms', 'https://admin:pwd@test.app.medicmobile.org/medic', undefined]);
-    expect(mocks.executeAction.args[1][0]).to.eq('compile-app-settings');
+    expectExecuteActionBehavior('convert-app-forms', undefined);
+    expect(mocks.executeAction.args[1][0].name).to.eq('compile-app-settings');
   });
 
   it('convert one form', async () => {
     const formName = 'form-name';
     await main([...normalArgv, '--local', 'convert-app-forms', '--', formName], {});
     expect(mocks.executeAction.callCount).to.deep.eq(1);
-    expect(mocks.executeAction.args[0]).to.deep.eq(['convert-app-forms', 'http://admin:pass@localhost:5988/medic', [formName]]);
+    expectExecuteActionBehavior('convert-app-forms', [formName]);
   });
 
   it('unsupported action', async () => {
@@ -120,15 +142,40 @@ describe('main', () => {
     expect(mocks.executeAction.called).to.be.false;
   });
 
-  describe('parseCouchUrl', () => {
-    const parseCouchUrl = main.__get__('parseCouchUrl');
-    it('basic', () =>
-      expect(parseCouchUrl('http://admin:pass@localhost:5988/medic').href).to.eq('http://admin:pass@localhost:5988/'));
-    
-    it('updates port', () =>
-      expect(parseCouchUrl('http://admin:pass@localhost:5984/medic').href).to.eq('http://admin:pass@localhost:5988/'));
+  describe('--archive', () => {
+    it('default actions', async () => {
+      await main([...normalArgv, '--archive', '--destination=foo'], {});
+      const executed = mocks.executeAction.args.map(args => args[0].name);
+      expect(executed).to.include('upload-app-settings');
+      expect(executed).to.not.include('delete-all-forms');
+    });
 
-    it('ignores path', () =>
-      expect(parseCouchUrl('http://admin:pass@localhost:5984/foo').href).to.eq('http://admin:pass@localhost:5988/'));
+    it('single action', async () => {
+      await main([...normalArgv, '--archive', '--destination=foo', 'upload-app-settings'], {});
+      expectExecuteActionBehavior('upload-app-settings');
+      expect(mocks.readline.keyInYN.callCount).to.eq(0);
+    });
+
+    it('requires destination', async () => {
+      const actual = await main([...normalArgv, '--archive', 'upload-app-settings'], {});
+      expect(actual).to.eq(-1);
+      expect(mocks.executeAction.called).to.be.false;
+    });
+  });
+
+  it('accept non-matching instance warning', async () => {
+    mocks.getApiUrl.returns('https://admin:pwd@url.app.medicmobile.org/medic');
+    mocks.readline.keyInYN.returns(true);
+    const actual = await main([...normalArgv, '---url=https://admin:pwd@url.app.medicmobile.org/']);
+    expect(mocks.readline.keyInYN.callCount).to.eq(1);
+    expect(actual).to.be.undefined;
+  });
+
+  it('reject non-matching instance warning', async () => {
+    mocks.getApiUrl.returns('https://admin:pwd@url.app.medicmobile.org/medic');
+    mocks.readline.keyInYN.returns(false);
+    const actual = await main([...normalArgv, '---url=https://admin:pwd@url.app.medicmobile.org/']);
+    expect(mocks.readline.keyInYN.callCount).to.eq(1);
+    expect(actual).to.eq(false);
   });
 });
