@@ -1,4 +1,3 @@
-const _ = require('lodash');
 const assert = require('chai').assert;
 
 const fs = require('../../src/lib/sync-fs');
@@ -7,19 +6,14 @@ describe('sync-fs', () => {
 
   describe('#withoutExtension()', () => {
 
-    _.forEach({
-      'person.xml': 'person',
-      'person.abc.xml': 'person.abc',
-    }, (expected, input) => {
+    [
+      ['person.xml', 'person'],
+      ['person.abc.xml', 'person.abc'],
+    ].forEach(([input, expected]) => {
 
       it(`should convert ${input} to ${expected}`, () => {
-
         assert.equal(fs.withoutExtension(input), expected);
-
       });
-
     });
-
   });
-
 });
